add a default include of our current package name
Its common in yocto to have a package_name.inc file to have common
shared data between versions. This is even more important given the fact
that cargo-bitbake generates the recipe for you so any overrides you'll
really want to put in that file so you don't have to continually rewrite
them.